Refactoring into modular structure with hierarchical CMakeLists.txt to enable building of unit tests.
