+++ src/common/CMAkeLists.txt
@@ -1,0 +1,1 @@
+add_subdirectory(error)
