+++ src/common/error/CMakeLists.txt
@@ -1,0 +1,6 @@
+add_library(error error.c)
+target_link_libraries(error PUBLIC pico_stdlib)
+
+target_include_directories(error PUBLIC
+  "${PROJECT_SOURCE_DIR}/src/common"
+)
