+++ src/common/macros.h
@@ -9,7 +9,11 @@
 #pragma once
 
 #include "error.h"
-#include "pico/printf.h"
+#if TEST == 1
+  #include <stdio.h>
+#else
+  #include "pico/printf.h"
+#endif
 
 /**
  * If this symbol is defined, we are configured to run a flight build.
