+++ src/common/typedefs.h
@@ -1,0 +1,4 @@
+typedef struct samwise_slate slate_t;
+typedef struct sched_state sched_state_t;
+typedef struct sched_task sched_task_t;
+typedef struct _rfm9x rfm9x_t;
