+++ src/drivers/CMakeLists.txt
@@ -1,0 +1,1 @@
+add_subdirectory(rfm9x)
