+++ src/init.c
@@ -8,9 +8,6 @@
  */
 
 #include "init.h"
-#include "macros.h"
-#include "pico/stdlib.h"
-#include "scheduler/scheduler.h"
 
 /**
  * Initialize all gpio pins to their default states.
