--- src/state_machine/states/init/init.h
@@ -1,6 +1,0 @@
-#pragma once
-
-#include "scheduler/scheduler.h"
-#include "slate.h"
-
-sched_state_t *init_get_next_state(slate_t *slate);
