--- src/state_machine/states/running/running.h
@@ -1,6 +1,0 @@
-#pragma once
-
-#include "scheduler/scheduler.h"
-#include "slate.h"
-
-sched_state_t *running_get_next_state(slate_t *slate);
