--- src/state_machine/tasks/tasks.h
@@ -1,24 +1,0 @@
-/**
- * @author  Joseph Shetaye
- * @date    2024-09-30
- *
- * Listing of all tasks. Tasks should not include this file - instead, states
- * should include this file to reference tasks in state definitions.
- */
-
-#pragma once
-
-#include "scheduler/scheduler.h"
-
-/*
- * Tasks are built from the tasks folder, which we cannot depend on to avoid
- * circularity.
- */
-
-// Extern -> assumes that these variables exist in another file, and accesses
-// them.
-
-// Each of these variables will be covered
-extern sched_task_t print_task;
-extern sched_task_t blink_task;
-extern sched_task_t radio_task;
