+++ src/states/CMakeLists.txt
@@ -1,0 +1,2 @@
+add_subdirectory(init)
+add_subdirectory(running)
