+++ src/states/running/CMakeLists.txt
@@ -1,0 +1,12 @@
+add_library(running_state running_state.c)
+
+# Specify targets to link compiled binaries
+target_link_libraries(running_state PUBLIC print_task blink_task radio_task)
+
+# Specify directories with relevant header files
+target_include_directories(running_state PUBLIC
+  "${PROJECT_SOURCE_DIR}/src/common"
+  "${PROJECT_SOURCE_DIR}/src/tasks/blink"
+  "${PROJECT_SOURCE_DIR}/src/tasks/print"
+  "${PROJECT_SOURCE_DIR}/src/tasks/radio"
+)
