+++ src/tasks/CMakeLists.txt
@@ -1,0 +1,3 @@
+add_subdirectory(blink)
+add_subdirectory(print)
+add_subdirectory(radio)
