+++ src/tasks/blink/blink_task.h
@@ -5,10 +5,14 @@
 
 #pragma once
 
-#include "scheduler/scheduler.h"
+#include "pico/stdlib.h"
+
+// The following ordering needs to be preserved
+#include "typedefs.h"
 #include "slate.h"
+#include "state_machine.h"
 
 void blink_task_init(slate_t *slate);
 void blink_task_dispatch(slate_t *slate);
 
-sched_task_t blink_task;
+extern sched_task_t blink_task;
