+++ src/tasks/print/print_task.h
@@ -6,10 +6,14 @@
 
 #pragma once
 
-#include "scheduler/scheduler.h"
+#include "macros.h"
+
+// The following ordering needs to be preserved
+#include "typedefs.h"
 #include "slate.h"
+#include "state_machine.h"
 
 void print_task_init(slate_t *slate);
 void print_task_dispatch(slate_t *slate);
 
-sched_task_t print_task;
+extern sched_task_t print_task;
