+++ src/tasks/print/test/print_test.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <string.h>
+
+#include "print_task.h"
