+++ src/tasks/radio/CMakeLists.txt
@@ -1,0 +1,10 @@
+add_library(radio_task radio_task.c)
+
+# Specify targets to link compiled binaries
+target_link_libraries(radio_task PUBLIC error pico_stdlib hardware_spi rfm9x)
+
+# Specify directories with relevant header files
+target_include_directories(radio_task PUBLIC
+  "${PROJECT_SOURCE_DIR}/src/common"
+  "${PROJECT_SOURCE_DIR}/src/drivers/rfm9x"
+)
