+++ src/tasks/radio/radio_task.h
@@ -5,8 +5,22 @@
 
 #pragma once
 
-#include "scheduler/scheduler.h"
+#include <string.h>
+
+#include "pico/stdlib.h"
+#include "pico/util/queue.h"
+#include "hardware/gpio.h"
+#include "hardware/spi.h"
+
+#include "macros.h"
+#include "pins.h"
+
+// The following ordering needs to be preserved
+#include "typedefs.h"
 #include "slate.h"
+#include "state_machine.h"
+
+#include "rfm9x.h"
 
 #define TX_QUEUE_SIZE 16
 #define RX_QUEUE_SIZE 16
@@ -24,4 +38,4 @@ typedef struct
 void radio_task_init(slate_t *slate);
 void radio_task_dispatch(slate_t *slate);
 
-sched_task_t radio_task;
+extern sched_task_t radio_task;
